fix #2244
